+++ src/app.ts
@@ -7,7 +7,7 @@ import { useSelector } from "react-redux";
 import loadable from "@loadable/component";
 import { Spinner } from "@blueprintjs/core";
 import "./styles/index.styl";
-import { useAppActions } from "./map-interface/reducers";
+import { useAppActions } from "~/map-interface/app-state";
 
 const _ColumnPage = loadable(import("./columns"));
 const ColumnPage = () => h(Suspense, { fallback: h(Spinner) }, h(_ColumnPage));
