+++ src/index.ts
@@ -8,8 +8,7 @@ FocusStyleManager.onlyShowFocusOnTabs();
 import { render } from "react-dom";
 import { Provider } from "react-redux";
 import { createStore, compose } from "redux";
-import reducers from "./map-interface/reducers";
-import { Action } from "./map-interface/reducers/actions";
+import reducerStack, { Action } from "./map-interface/app-state";
 import App from "./app";
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
@@ -18,7 +17,7 @@ type AppState = any;
 
 // Create the data store
 let store = createStore<AppState, Action, any, any>(
-  reducers,
+  reducerStack,
   composeEnhancers()
 );
