+++ src/map-interface/app-state/actions/map-state.ts
