+++ src/map-interface/app-state/handlers/fetch.ts
