+++ src/map-interface/app-state/handlers/filters.ts
