+++ src/map-interface/app-state/handlers/index.ts
