+++ src/map-interface/app-state/helpers/hash-string.ts
