+++ src/map-interface/app-state/index.ts
@@ -49,5 +49,6 @@ function overallReducer(state, action: Action) {
 const newReducer = reduceReducers(overallReducer, reducers);
 
 export default newReducer;
+export * from "./actions";
 export * from "./hooks";
 export * from "./menu";
