+++ src/map-interface/app-state/menu.ts
