+++ src/map-interface/components/Searchbar.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Navbar, Button, InputGroup, Card, Spinner } from "@blueprintjs/core";
 import h from "@macrostrat/hyper";
-import { useAppActions, useSearchState } from "../reducers";
+import { useAppActions, useSearchState } from "../app-state";
 import { useSelector } from "react-redux";
 import { useEffect } from "react";
 import { SubtleFilterText } from "./filters-panel";
