+++ src/map-interface/map-page/index.ts
@@ -9,8 +9,7 @@ import ElevationChart from "../components/elevation-chart";
 import { ButtonGroup, Button, Spinner } from "@blueprintjs/core";
 import { useSelector, useDispatch } from "react-redux";
 import loadable from "@loadable/component";
-import { useSearchState } from "../reducers";
-import { MapBackend } from "../reducers/actions";
+import { useSearchState, MapBackend } from "../app-state";
 import styles from "./main.module.styl";
 
 const h = hyper.styled(styles);
